Sync Update

+++ plugins/webgui_order/index.js
@@ -1,7 +1,36 @@
 const knex = appRequire('init/knex').knex;
+const config = appRequire('services/config').all();
 
 const getOrders = async () => {
   return knex('webgui_order').where({});
+};
+
+const getOrdersAndAccountNumber = async () => {
+  const orders = await knex('webgui_order').select([
+    'webgui_order.id as id',
+    'webgui_order.baseId as baseId',
+    'webgui_order.name as name',
+    'webgui_order.shortComment as shortComment',
+    'webgui_order.comment as comment',
+    'webgui_order.type as type',
+    'webgui_order.cycle as cycle',
+    'webgui_order.alipay as alipay',
+    'webgui_order.paypal as paypal',
+    'webgui_order.flow as flow',
+    'webgui_order.refTime as refTime',
+    'webgui_order.server as server',
+    'webgui_order.autoRemove as autoRemove',
+    'webgui_order.autoRemoveDelay as autoRemoveDelay',
+    'webgui_order.portRange as portRange',
+    'webgui_order.multiServerFlow as multiServerFlow',
+    'webgui_order.changeOrderType as changeOrderType',
+    'webgui_order.autoRemove as autoRemove',
+    knex.raw('count(account_plugin.id) as accountNumber'),
+  ])
+  .leftJoin('account_plugin', 'account_plugin.orderId', 'webgui_order.id')
+  .groupBy('webgui_order.id')
+  .orderBy('webgui_order.name', 'ASC');
+  return orders;
 };
 
 const getOneOrder = async orderId => {
@@ -20,8 +49,10 @@ const getOneOrderByAccountId = async accountId => {
 };
 
 const newOrder = async data => {
-  await knex('webgui_order').insert({
+  const [ id ] = await knex('webgui_order').insert({
+    baseId: data.baseId,
     name: data.name,
+    shortComment: data.shortComment,
     comment: data.comment,
     type: data.type,
     cycle: data.cycle,
@@ -31,15 +62,20 @@ const newOrder = async data => {
     refTime: data.refTime,
     server: data.server ? JSON.stringify(data.server) : null,
     autoRemove: data.autoRemove,
+    autoRemoveDelay: data.autoRemoveDelay,
+    portRange: data.portRange,
     multiServerFlow: data.multiServerFlow,
     changeOrderType: data.changeOrderType,
+    active: data.active,
   });
-  return;
+  return id;
 };
 
 const editOrder = async data => {
   await knex('webgui_order').update({
+    baseId: data.baseId,
     name: data.name,
+    shortComment: data.shortComment,
     comment: data.comment,
     type: data.type,
     cycle: data.cycle,
@@ -49,8 +85,11 @@ const editOrder = async data => {
     refTime: data.refTime,
     server: data.server ? JSON.stringify(data.server) : null,
     autoRemove: data.autoRemove,
+    autoRemoveDelay: data.autoRemoveDelay,
+    portRange: data.portRange,
     multiServerFlow: data.multiServerFlow,
     changeOrderType: data.changeOrderType,
+    active: data.active,
   }).where({
     id: data.id,
   });
@@ -58,15 +97,24 @@ const editOrder = async data => {
 };
 
 const deleteOrder = async orderId => {
-  const hasAccount = await knex('account_plugin').where({ orderId });
-  if(hasAccount.length) { return Promise.reject('account with this order exists'); }
-  const hasGiftcard = await knex('giftcard').where({ orderType: orderId, status: 'AVAILABLE' });
-  if(hasGiftcard.length) { return Promise.reject('giftcard with this order exists'); }
-  await knex('webgui_order').delete().where({ id: orderId });
+  const orderInfo = await knex('webgui_order').where({ id: orderId }).then(s => s[0]);
+  if(orderInfo.baseId) {
+    await knex('webgui_order').delete().where({ id: orderId });
+  } else {
+    const hasAccount = await knex('account_plugin').where({ orderId });
+    if(hasAccount.length) { return Promise.reject('account with this order exists'); }
+    const isGiftCardOn = config.plugins.giftcard && config.plugins.giftcard.use;
+    const hasGiftcard = isGiftCardOn ? await knex('giftcard').where({ orderType: orderId, status: 'AVAILABLE' }) : [];
+    if(hasGiftcard.length) { return Promise.reject('giftcard with this order exists'); }
+    const hasFlowPackOrder = await knex('webgui_order').where({ baseId: orderId });
+    if(hasFlowPackOrder.length) { return Promise.reject('flowpack order exists'); }
+    await knex('webgui_order').delete().where({ id: orderId });
+  }
   return;
 };
 
 exports.getOrders = getOrders;
+exports.getOrdersAndAccountNumber = getOrdersAndAccountNumber;
 exports.getOneOrder = getOneOrder;
 exports.newOrder = newOrder;
 exports.editOrder = editOrder;
